refactor: utils 패키지 제거
- 일부 클래스는 common 패키지로 이동

Related to : #251
