+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/common/bindingadapter/ViewPagerBindingAdapter.kt
@@ -1,4 +1,4 @@
-package com.emmsale.presentation.utils.bindingadapter
+package com.emmsale.presentation.common.bindingadapter
 
 import androidx.databinding.BindingAdapter
 import androidx.viewpager2.widget.ViewPager2
