+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/eventdetail/uistate/EventDetailUiState.kt
@@ -1,7 +1,10 @@
 package com.emmsale.presentation.ui.eventdetail.uistate
 
 import com.emmsale.data.eventdetail.EventDetail
-import com.emmsale.presentation.utils.DateStringFormat
+import java.time.LocalDate
+import java.time.LocalDateTime
+import java.time.LocalTime
+import java.time.format.DateTimeFormatter
 
 sealed class EventDetailUiState {
     data class Success(
@@ -26,13 +29,25 @@ sealed class EventDetailUiState {
                     name = name,
                     status = status,
                     location = location,
-                    startDate = DateStringFormat.getGeneralDateFormat(startDate),
-                    endDate = DateStringFormat.getGeneralDateFormat(endDate),
+                    startDate = getGeneralDateFormat(startDate),
+                    endDate = getGeneralDateFormat(endDate),
                     informationUrl = informationUrl,
                     tags = tags,
                     imageUrl = imageUrl,
                 )
             }
+        }
+
+        private fun getGeneralDateFormat(dateString: String): String {
+            val formatter = DateTimeFormatter.ofPattern("yyyy:MM:dd:HH:mm:ss")
+            val dateTime = LocalDateTime.parse(dateString, formatter)
+
+            val targetDate = LocalDate.of(dateTime.year, dateTime.month, dateTime.dayOfMonth)
+            val targetTime = LocalTime.of(dateTime.hour, dateTime.minute)
+            val targetDateTime = LocalDateTime.of(targetDate, targetTime)
+
+            val resultFormatter = DateTimeFormatter.ofPattern("yyyy.M.d HH:mm")
+            return targetDateTime.format(resultFormatter)
         }
     }
 }
