--- android/2023-emmsale/app/src/main/java/com/emmsale/presentation/utils/builder/UriBuilder.kt
@@ -1,5 +1,0 @@
-package com.emmsale.presentation.utils.builder
-
-import android.net.Uri
-
-fun uri(block: Uri.Builder.() -> Unit): Uri = Uri.Builder().apply(block).build()
