+++ android/2023-emmsale/app/src/main/res/values/strings.xml
@@ -58,8 +58,6 @@
     <string name="myprofile_clubslabel">동아리 활동</string>
     <string name="myprofile_eventslabel">참여 행사</string>
 
-    <string name="event_start_duration_format">기간 : %s</string>
-    <string name="event_start_end_duration_format">기간 : %s - %s</string>
     <string name="event_poster_description">event poster</string>
     <string name="event_filter">필터 설정</string>
     <string name="event_count_format">%,d</string>
